73 - used ctx->request to *lock down* mutation resolvers to logged in users

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -114,32 +114,63 @@ async deletePost(parent, args, { prisma, request }, info){
         
     }, info)
 },
-async updatePost(parent, args, { prisma }, info){
-return prisma.mutation.updatePost({
-    where:{
-        id: args.id
-    },
-    data: args.data
+async updatePost(parent, args, { prisma, request }, info){
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+        id: args.id,
+        author: {  
+            id: userId
+        }
+    })
+    if (!postExists) {
+        throw new Error('Unable to update')
+    }
+
+
+
+    return prisma.mutation.updatePost({
+    
+        where: {
+            id: args.id
+        },
+        data: args.data
+
 }, info)
 }, 
-async createComment(parent, args, { prisma }, info){
+async createComment(parent, args, { prisma, request }, info){
+    const userId = getUserId(request)
+
+
+
+    
     return prisma.mutation.createComment({
         data:{
             text: args.data.text,
-        author:{
-            connect:{
-                id: args.data.author
-            }
+            author: {
+                connect:{
+                id: userId
+            }  
+          
         },
         post:{
             connect: {
                 id: args.data.post
             }
         }
-    }
+        }
     }, info)
 },
-deleteComment(parent, args, { prisma }, info){
+async deleteComment(parent, args, { prisma, request }, info){
+const userId = getUserId(request)
+const commentExists = await prisma.exists.Comment({
+    id: args.id,
+    author:{
+        id: userId
+    }
+})
+ if (!commentExists){
+     throw new Error('Unable to delete comment')
+ }
 return prisma.mutation.deleteComment({
     where: {
         id: args.id
@@ -147,8 +178,18 @@ return prisma.mutation.deleteComment({
 }, info)
 
 },
-updateComment (parent, args, { prisma }, info){
-return prisma.mutation.updateComment({
+async updateComment (parent, args, { prisma, request }, info){
+    const userId = getUserId(request)
+    const commentExists = await prisma.exists.Comment({
+        id: args.id,
+        author: {
+            id: userId
+        }
+    })
+    if (!commentExists){
+        throw new Error ('Unable to delete comment')
+    }
+    return prisma.mutation.updateComment({
     where:{
         id: args.id
     },
